+++ src/main/java/dao/coupon/CouponDaoImpl.java
@@ -39,8 +39,8 @@ public class CouponDaoImpl implements CouponDao {
     }
 
     @Override
-    public Coupon getOne( long id ) {
-        Coupon coupon= null;
+    public Coupon getOne( int id ) {
+        Coupon coupon = null;
         Connection connection = pool.getConnection();
         try {
             PreparedStatement st = connection.prepareStatement( "SELECT * FROM coupons WHERE coupon_id = ?" );
@@ -49,16 +49,16 @@ public class CouponDaoImpl implements CouponDao {
 
             if ( rs.next() ) {
                 String title = rs.getString( 2 );
-                Date startDate = Date.valueOf(rs.getString( 3 ));
-                Date endDate = Date.valueOf(rs.getString( 4 ));
-                int amount = Integer.parseInt(rs.getString( 5 ));
-                CouponType type = CouponType.valueOf(rs.getString( 6 ));
+                Date startDate = Date.valueOf( rs.getString( 3 ) );
+                Date endDate = Date.valueOf( rs.getString( 4 ) );
+                int amount = Integer.parseInt( rs.getString( 5 ) );
+                CouponType type = CouponType.valueOf( rs.getString( 6 ) );
                 String message = rs.getString( 7 );
-                double price = Double.parseDouble(rs.getString( 8 ));
+                double price = Double.parseDouble( rs.getString( 8 ) );
                 String image = rs.getString( 9 );
 
 
-                coupon = new Coupon( id, title,startDate,endDate,amount,type,message,price,image);
+                coupon = new Coupon( id, title, startDate, endDate, amount, type, message, price, image );
             }
         } catch ( SQLException e ) {
 
@@ -73,7 +73,7 @@ public class CouponDaoImpl implements CouponDao {
     public int create( Coupon coupon ) {
         try {
             Statement st = connection.createStatement();
-            String query = String.format( "INSERT INTO coupons (title,startDate endDate,amount,type,message,price,image) VALUES ('%s', '%s', '%s','%s','%s','%s','%s','%')", coupon.getTitle(), coupon.getStartDate(),coupon.getEndDate(),coupon.getAmount(),coupon.getType(),coupon.getMessage(),coupon.getPrice(),coupon.getImage());
+            String query = String.format( "INSERT INTO coupons (title,startDate endDate,amount,type,message,price,image) VALUES ('%s', '%s', '%s','%s','%s','%s','%s','%')", coupon.getTitle(), coupon.getStartDate(), coupon.getEndDate(), coupon.getAmount(), coupon.getType(), coupon.getMessage(), coupon.getPrice(), coupon.getImage() );
 
 
             int result = st.executeUpdate( query, Statement.RETURN_GENERATED_KEYS );
@@ -90,7 +90,7 @@ public class CouponDaoImpl implements CouponDao {
             }
         } catch ( SQLException e ) {
             e.printStackTrace();
-        } catch (UniqueValueException e) {
+        } catch ( UniqueValueException e ) {
             e.printStackTrace();
         } finally {
             pool.returnConnection( connection );
@@ -103,7 +103,7 @@ public class CouponDaoImpl implements CouponDao {
     public void update( Coupon coupon ) throws SqlServerException {
         try {
             Statement st = connection.createStatement();
-            String sql = String.format( "UPDATE coupons SET title = %s, startDate = %s, endDate = %s, amount = %s, type = %s,message = %s,price = %s,image = %s, WHERE id = %d", coupon.getTitle(),coupon.getStartDate(),coupon.getEndDate(),coupon.getAmount(),coupon.getType(),coupon.getMessage(),coupon.getPrice(),coupon.getImage(),coupon.getId());
+            String sql = String.format( "UPDATE coupons SET title = %s, startDate = %s, endDate = %s, amount = %s, type = %s,message = %s,price = %s,image = %s, WHERE id = %d", coupon.getTitle(), coupon.getStartDate(), coupon.getEndDate(), coupon.getAmount(), coupon.getType(), coupon.getMessage(), coupon.getPrice(), coupon.getImage(), coupon.getId() );
             int result = st.executeUpdate( sql );
 
             if ( result == 0 ) {
@@ -118,9 +118,8 @@ public class CouponDaoImpl implements CouponDao {
     }
 
 
-
     @Override
-    public void delete( long id ) throws SQLException {
+    public void delete( int id ) throws SQLException {
         Connection connection = pool.getConnection();
         try {
             connection.setAutoCommit( false );
